requirements.txt
mcp
chromadb
openai
python-dotenv
PyYAML
